+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ventura-ignite-ui/react
 
+## 2.1.0
+
+### Minor Changes
+
+- Adiciona novo componente de tooltip
+
 ## 2.0.1
 
 ### Patch Changes