+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ventura-ignite-ui/tokens
 
+## 1.1.1
+
+### Patch Changes
+
+- Adiciona novo font family Inter
+
 ## 1.1.0
 
 ### Minor Changes